docs(ComboboxSimple): add dedicated mdx docs for ComboboxSimple
