+++ apps/docs/stories/combobox-primitive.mdx
@@ -1,0 +1,182 @@
+import { Meta, Controls, Primary } from '@storybook/addon-docs/blocks';
+import * as ComboboxPrimitiveStories from './combobox-primitive.stories';
+import * as ComboboxTriggerStories from './combobox-trigger.stories';
+import * as ComboboxContentStories from './combobox-content.stories';
+import * as ComboboxCommandStories from './combobox-command.stories';
+import * as ComboboxInputStories from './combobox-input.stories';
+import * as ComboboxListStories from './combobox-list.stories';
+import * as ComboboxEmptyStories from './combobox-empty.stories';
+import * as ComboboxGroupStories from './combobox-group.stories';
+import * as ComboboxItemStories from './combobox-item.stories';
+import * as ComboboxLoadingStories from './combobox-loading.stories';
+import * as ComboboxSeparatorStories from './combobox-separator.stories';
+
+<Meta of={ComboboxPrimitiveStories} />
+
+# Combobox
+
+Low-level building blocks for creating custom combobox experiences. Compose `Combobox`, `ComboboxTrigger`, `ComboboxContent`, `ComboboxCommand`, `ComboboxInput`, `ComboboxList`, and `ComboboxItem` for full control over behaviour and layout.
+
+For a ready-to-use preset with filtering, multi-select, and grouped items, see [ComboboxSimple](?path=/docs/composed-components-comboboxsimple--docs).
+
+## Composition example
+
+```tsx
+import {
+	Combobox,
+	ComboboxTrigger,
+	ComboboxContent,
+	ComboboxCommand,
+	ComboboxInput,
+	ComboboxList,
+	ComboboxItem,
+	ComboboxEmpty,
+} from '@signozhq/ui';
+import { useState } from 'react';
+
+const frameworks = [
+	{ value: 'react', label: 'React' },
+	{ value: 'vue', label: 'Vue' },
+];
+
+export default function MyComponent() {
+	const [value, setValue] = useState('');
+	const [open, setOpen] = useState(false);
+	const selectedLabel = frameworks.find((f) => f.value === value)?.label ?? '';
+
+	return (
+		<Combobox open={open} onOpenChange={setOpen}>
+			<ComboboxTrigger placeholder="Select a framework..." value={selectedLabel} />
+			<ComboboxContent>
+				<ComboboxCommand>
+					<ComboboxInput placeholder="Search frameworks..." />
+					<ComboboxList>
+						<ComboboxItem
+							value="react"
+							onSelect={() => {
+								setValue('react');
+								setOpen(false);
+							}}
+							isSelected={value === 'react'}
+						>
+							React
+						</ComboboxItem>
+						<ComboboxItem
+							value="vue"
+							onSelect={() => {
+								setValue('vue');
+								setOpen(false);
+							}}
+							isSelected={value === 'vue'}
+						>
+							Vue
+						</ComboboxItem>
+						<ComboboxEmpty>No results.</ComboboxEmpty>
+					</ComboboxList>
+				</ComboboxCommand>
+			</ComboboxContent>
+		</Combobox>
+	);
+}
+```
+
+### Combobox inside a popover
+
+When placing inside a popover, set `withPortal={false}` on `ComboboxContent` to avoid z-index and stacking issues:
+
+```tsx
+import {
+	Popover,
+	PopoverTrigger,
+	PopoverContent,
+	Combobox,
+	ComboboxTrigger,
+	ComboboxContent,
+	ComboboxCommand,
+	ComboboxInput,
+	ComboboxList,
+} from '@signozhq/ui';
+
+<Popover>
+	<PopoverTrigger asChild>
+		<Button variant="outline">Open</Button>
+	</PopoverTrigger>
+	<PopoverContent>
+		<Combobox open={open} onOpenChange={setOpen}>
+			<ComboboxTrigger placeholder="Select..." value={selectedLabel} />
+			<ComboboxContent withPortal={false}>
+				<ComboboxCommand>
+					<ComboboxInput placeholder="Search..." />
+					<ComboboxList>{/* items */}</ComboboxList>
+				</ComboboxCommand>
+			</ComboboxContent>
+		</Combobox>
+	</PopoverContent>
+</Popover>
+```
+
+### Loading state (primitive composition)
+
+Use `ComboboxLoading` inside `ComboboxList` when fetching options asynchronously:
+
+```tsx
+<ComboboxList>
+	{isLoading ? (
+		<ComboboxLoading>Loading options...</ComboboxLoading>
+	) : (
+		items.map((item) => (
+			<ComboboxItem key={item.value} value={item.value}>
+				{item.label}
+			</ComboboxItem>
+		))
+	)}
+</ComboboxList>
+```
+
+<Primary />
+
+## Combobox Props
+
+<Controls of={ComboboxPrimitiveStories.Default} />
+
+## Subcomponents
+
+### ComboboxTrigger
+
+<Controls of={ComboboxTriggerStories.Default} />
+
+### ComboboxContent
+
+<Controls of={ComboboxContentStories.Default} />
+
+### ComboboxCommand
+
+<Controls of={ComboboxCommandStories.Default} />
+
+### ComboboxInput
+
+<Controls of={ComboboxInputStories.Default} />
+
+### ComboboxList
+
+<Controls of={ComboboxListStories.Default} />
+
+### ComboboxEmpty
+
+<Controls of={ComboboxEmptyStories.Default} />
+
+### ComboboxGroup
+
+<Controls of={ComboboxGroupStories.Default} />
+
+### ComboboxItem
+
+<Controls of={ComboboxItemStories.Default} />
+
+### ComboboxLoading
+
+<Controls of={ComboboxLoadingStories.Default} />
+
+### ComboboxSeparator
+
+<Controls of={ComboboxSeparatorStories.Default} />
